Legacy_Code_Conv/Imaging/Modernization/requirements.txt
# API Framework
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Storage
boto3>=1.29.3
aioboto3>=12.1.0
aiofiles>=23.2.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Image Processing
Pillow>=10.1.0
python-magic>=0.4.27

# Caching
redis>=5.0.1

# Monitoring
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
flake8>=6.1.0
